// projects/admin-jobroad/src/app/dto/job/job.res.dto.ts
export interface JobResDto {
  id: string;
  jobName: string;
  companyId : string;
  companyName: string;
  address: string;
  startDate: string;
  endDate: string;
  description: string;
  expectedSalaryMin: string;
  expectedSalaryMax: string;
  employmentTypeId : string;
  employementTypeName: string;
  fileId: string;
  companyPhotoId: string;
  hrId: string;
  picId: string;
  createdBy: string
}